requirements.txt
requests~=2.32.3
python-dotenv~=1.0.1
cachetools~=5.5.2
argparse
Flask